Moved osmclient directory one level up
