+++ src/osmclient/common/exceptions.py
