+++ src/osmclient/common/http.py
