+++ src/osmclient/common/test/test_utils.py
