+++ src/osmclient/common/utils.py
