+++ src/osmclient/scripts/osm.py
