+++ src/osmclient/sol005/nsi.py
@@ -18,10 +18,10 @@
 OSM NSI (Network Slice Instance) API handling
 """
 
-from tngsdk.osmclient.common import utils
-from tngsdk.osmclient.common import wait as WaitForStatus
-from tngsdk.osmclient.common.exceptions import ClientException
-from tngsdk.osmclient.common.exceptions import NotFound
+from osmclient.common import utils
+from osmclient.common import wait as WaitForStatus
+from osmclient.common.exceptions import ClientException
+from osmclient.common.exceptions import NotFound
 import yaml
 import json
