+++ src/osmclient/sol005/package.py
@@ -20,9 +20,9 @@ OSM package API handling
 
 #from os import stat
 #from os.path import basename
-from tngsdk.osmclient.common.exceptions import ClientException
-from tngsdk.osmclient.common.exceptions import NotFound
-from tngsdk.osmclient.common import utils
+from osmclient.common.exceptions import ClientException
+from osmclient.common.exceptions import NotFound
+from osmclient.common import utils
 import json
