+++ src/osmclient/sol005/role.py
@@ -20,9 +20,9 @@
 OSM role mgmt API
 """
 
-from tngsdk.osmclient.common import utils
-from tngsdk.osmclient.common.exceptions import ClientException
-from tngsdk.osmclient.common.exceptions import NotFound
+from osmclient.common import utils
+from osmclient.common.exceptions import ClientException
+from osmclient.common.exceptions import NotFound
 import json
 import yaml
