+++ src/osmclient/sol005/user.py
@@ -19,9 +19,9 @@
 OSM user mgmt API
 """
 
-from tngsdk.osmclient.common import utils
-from tngsdk.osmclient.common.exceptions import ClientException
-from tngsdk.osmclient.common.exceptions import NotFound
+from osmclient.common import utils
+from osmclient.common.exceptions import ClientException
+from osmclient.common.exceptions import NotFound
 import json
