+++ src/osmclient/sol005/vnfd.py
@@ -15,64 +15,65 @@
 #    under the License.
 
 """
-OSM NST (Network Slice Template) API handling
+OSM vnfd API handling
 """
 
-from tngsdk.osmclient.common.exceptions import NotFound
-from tngsdk.osmclient.common.exceptions import ClientException
-from tngsdk.osmclient.common import utils
+from osmclient.common.exceptions import NotFound
+from osmclient.common.exceptions import ClientException
+from osmclient.common import utils
 import json
 import magic
+from os.path import basename
 #from os import stat
-#from os.path import basename
 
-class Nst(object):
+
+class Vnfd(object):
 
     def __init__(self, http=None, client=None):
         self._http = http
         self._client = client
-        self._apiName = '/nst'
+        self._apiName = '/vnfpkgm'
         self._apiVersion = '/v1'
-        self._apiResource = '/netslice_templates'
+        self._apiResource = '/vnf_packages'
         self._apiBase = '{}{}{}'.format(self._apiName,
                                         self._apiVersion, self._apiResource)
+        #self._apiBase='/vnfds'
 
     def list(self, filter=None):
         filter_string = ''
         if filter:
             filter_string = '?{}'.format(filter)
-        resp = self._http.get_cmd('{}{}'.format(self._apiBase, filter_string))
-        #print yaml.safe_dump(resp)
+        resp = self._http.get_cmd('{}{}'.format(self._apiBase,filter_string))
         if resp:
             return resp
         return list()
 
     def get(self, name):
         if utils.validate_uuid4(name):
-            for nst in self.list():
-                if name == nst['_id']:
-                    return nst
+            for vnfd in self.list():
+                if name == vnfd['_id']:
+                    return vnfd
         else:
-            for nst in self.list():
-                if 'name' in nst and name == nst['name']:
-                    return nst
-        raise NotFound("nst {} not found".format(name))
+            for vnfd in self.list():
+                if 'name' in vnfd and name == vnfd['name']:
+                    return vnfd
+        raise NotFound("vnfd {} not found".format(name))
 
     def get_individual(self, name):
-        nst = self.get(name)
-        # It is redundant, since the previous one already gets the whole nstinfo
+        vnfd = self.get(name)
+        # It is redundant, since the previous one already gets the whole vnfpkginfo
         # The only difference is that a different primitive is exercised
-        resp = self._http.get_cmd('{}/{}'.format(self._apiBase, nst['_id']))
+        resp = self._http.get_cmd('{}/{}'.format(self._apiBase, vnfd['_id']))
         #print yaml.safe_dump(resp)
         if resp:
             return resp
-        raise NotFound("nst {} not found".format(name))
+        raise NotFound("vnfd {} not found".format(name))
 
     def get_thing(self, name, thing, filename):
-        nst = self.get(name)
+        vnfd = self.get(name)
         headers = self._client._headers
         headers['Accept'] = 'application/binary'
-        http_code, resp = self._http.get2_cmd('{}/{}/{}'.format(self._apiBase, nst['_id'], thing))
+        http_code, resp = self._http.get2_cmd('{}/{}/{}'.format(self._apiBase, vnfd['_id'], thing))
         #print 'HTTP CODE: {}'.format(http_code)
         #print 'RESP: {}'.format(resp)
         if http_code in (200, 201, 202, 204):
@@ -89,21 +90,21 @@ class Nst(object):
             raise ClientException("failed to get {} from {} - {}".format(thing, name, msg))
 
     def get_descriptor(self, name, filename):
-        self.get_thing(name, 'nst', filename)
+        self.get_thing(name, 'vnfd', filename)
 
     def get_package(self, name, filename):
-        self.get_thing(name, 'nst_content', filename)
+        self.get_thing(name, 'package_content', filename)
 
     def get_artifact(self, name, artifact, filename):
         self.get_thing(name, 'artifacts/{}'.format(artifact), filename)
 
     def delete(self, name, force=False):
-        nst = self.get(name)
+        vnfd = self.get(name)
         querystring = ''
         if force:
             querystring = '?FORCE=True'
         http_code, resp = self._http.delete_cmd('{}/{}{}'.format(self._apiBase,
-                                         nst['_id'], querystring))
+                                         vnfd['_id'], querystring))
         #print 'HTTP CODE: {}'.format(http_code)
         #print 'RESP: {}'.format(resp)
         if http_code == 202:
@@ -114,10 +115,10 @@ class Nst(object):
             msg = ""
             if resp:
                 try:
-                    resp = json.loads(resp)
+                    msg = json.loads(resp)
                 except ValueError:
                     msg = resp
-            raise ClientException("failed to delete nst {} - {}".format(name, msg))
+            raise ClientException("failed to delete vnfd {} - {}".format(name, msg))
 
     def create(self, filename, overwrite=None, update_endpoint=None):
         mime_type = magic.from_file(filename, mime=True)
@@ -125,8 +126,9 @@ class Nst(object):
             raise ClientException(
                      "failed to guess MIME type for file '{}'".format(filename))
         headers= self._client._headers
-        if mime_type in ['application/yaml', 'text/plain']:
-            headers['Content-Type'] = 'application/yaml'
+        headers['Content-Filename'] = basename(filename)
+        if mime_type in ['application/yaml', 'text/plain', 'application/json']:
+            headers['Content-Type'] = 'text/plain'
         elif mime_type in ['application/gzip', 'application/x-gzip']:
             headers['Content-Type'] = 'application/gzip'
             #headers['Content-Type'] = 'application/binary'
@@ -149,20 +151,22 @@ class Nst(object):
             ow_string = ''
             if overwrite:
                 ow_string = '?{}'.format(overwrite)
-            self._apiResource = '/netslice_templates_content'
+            self._apiResource = '/vnf_packages_content'
             self._apiBase = '{}{}{}'.format(self._apiName,
                                             self._apiVersion, self._apiResource)
             endpoint = '{}{}'.format(self._apiBase,ow_string)
             http_code, resp = self._http.post_cmd(endpoint=endpoint, filename=filename)
         #print 'HTTP CODE: {}'.format(http_code)
         #print 'RESP: {}'.format(resp)
-        if http_code in (200, 201, 202, 204):
+        if http_code in (200, 201, 202):
             if resp:
                 resp = json.loads(resp)
             if not resp or 'id' not in resp:
-                raise ClientException('unexpected response from server - {}'.format(
+                raise ClientException('unexpected response from server: '.format(
                                       resp))
-            print(resp['id'])
+            return resp['id']
+        elif http_code == 204:
+            print('Updated')
         else:
             msg = "Error {}".format(http_code)
             if resp:
@@ -170,10 +174,10 @@ class Nst(object):
                     msg = "{} - {}".format(msg, json.loads(resp))
                 except ValueError:
                     msg = "{} - {}".format(msg, resp)
-            raise ClientException("failed to create/update nst - {}".format(msg))
+            raise ClientException("failed to create/update vnfd - {}".format(msg))
 
     def update(self, name, filename):
-        nst = self.get(name)
-        endpoint = '{}/{}/nst_content'.format(self._apiBase, nst['_id'])
+        vnfd = self.get(name)
+        endpoint = '{}/{}/package_content'.format(self._apiBase, vnfd['_id'])
         self.create(filename=filename, update_endpoint=endpoint)
