+++ src/osmclient/v1/tests/test_ns.py
