+++ src/osmclient/v1/tests/test_vnf.py
