+++ src/osmclient/v1/tests/test_vnfd.py
