+++ src/osmclient/v1/vim.py
@@ -18,8 +18,8 @@
 OSM vim API handling
 """
 
-from tngsdk.osmclient.common.exceptions import ClientException
-from tngsdk.osmclient.common.exceptions import NotFound
+from osmclient.common.exceptions import ClientException
+from osmclient.common.exceptions import NotFound
 import yaml 
 import time
