+++ src/osmclient/v1/vnf.py
@@ -18,7 +18,7 @@
 OSM vnf API handling
 """
 
-from tngsdk.osmclient.common.exceptions import NotFound
+from osmclient.common.exceptions import NotFound
 
 
 class Vnf(object):
